fix build fail(UISelectExtended initialization issues)
